My Solution (Time Up for checkpoint challenge)
