+++ school.rb
@@ -1,0 +1,16 @@
+
+class School
+  attr_reader :age, :salary, :phase, :performance_rating, :target_raise
+  attr_accessor :name
+
+  def initialize(options={})
+    @age = options.fetch(:age, 0)
+    @name = options.fetch(:name, "")
+    @phase = 1
+  end
+
+  def offer_high_five
+    "High five!"
+  end
+
+end
